Compile OK

+++ libft/srcs/ft_strnequ.c
@@ -6,7 +6,7 @@
 /*   By: jehpark <jehpark@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/26 19:58:24 by jehpark           #+#    #+#             */
-/*   Updated: 2021/04/27 08:07:57 by jehpark          ###   ########.fr       */
+/*   Updated: 2021/04/27 10:39:45 by jehpark          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,13 +18,13 @@ int		ft_strnequ(char const *s1, char const *s2, size_t n)
 	int	ans;
 
 	i = 0;
-	while (*s1 && *s1 == *s2 && i < n)
+	while (*s1 && *s1 == *s2 && (size_t)i < n)
 	{
 		s1++;
 		s2++;
 		i++;
 	}
-	if (i == n)
+	if ((size_t)i == n)
 	{
 		s1--;
 		s2--;
